Fix completion key issue with CreateIoCompletionPort
Addressed a bug where completion key arguments were incorrectly typed. They're now either a manually defined uint or a reference to the watched directory handle that we cast to a uint. As a result, the program no longer uses the ID array (overlapped, the watched directory handle, and the watched directory) as a completion key. It might be useful to figure out how to do that in the future though as that would come in handy for watching multiple directories.