Refactored process execution metadata
The metadata used to track executed processes was moved out of a struct and into individual variables. It wasn't clear to me what the benefits of using a struct for these attributes were. It might be worth revisiting this later on and grouping attributes in a more logical fashion. I.e. all read and write handles together, etc. At the very least this would make it easier to reset these attributes in one go.